LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/alignment/flex-align-baseline-flex-003-expected.txt


PASS .target > * 1
PASS .target > * 2
FAIL .target > * 3 assert_equals:
<div data-offset-y="45"><span></span></div>
offsetTop expected 45 but got 15
PASS .target > * 4
FAIL .target > * 5 assert_equals:
<div data-offset-y="55"><span></span></div>
offsetTop expected 55 but got 115
PASS .target > * 6
FAIL .target > * 7 assert_equals:
<div data-offset-y="55"><span></span></div>
offsetTop expected 55 but got 85
PASS .target > * 8
FAIL .target > * 9 assert_equals:
<div data-offset-y="115"><span></span></div>
offsetTop expected 115 but got 85
PASS .target > * 10
PASS .target > * 11
PASS .target > * 12
FAIL .target > * 13 assert_equals:
<div data-offset-y="115"><span></span></div>
offsetTop expected 115 but got 25
PASS .target > * 14
FAIL .target > * 15 assert_equals:
<div data-offset-y="115"><span></span></div>
offsetTop expected 115 but got 55
PASS .target > * 16
